Тест фронтенда на Leptos, чуть обновлённый README.md.
