+++ .taurignore
@@ -1,3 +1,2 @@
 /src
-/public
 /Cargo.toml
