+++ Cargo.toml
@@ -4,11 +4,11 @@ version = "2.0.0-beta.0"
 edition = "2021"
 
 [dependencies]
-sycamore = { version = "0.8.2", features = ["suspense"] }
+leptos = { version = "0.6.15", features = ["csr"] }
 wasm-bindgen = "0.2.95"
 wasm-bindgen-futures = "0.4.45"
 js-sys = "0.3.72"
-serde = { version = "1.0.210", features = ["derive"] }
+serde = { version = "1.0.214", features = ["derive"] }
 serde-wasm-bindgen = "0.6.5"
 console_error_panic_hook = "0.1.7"
