+++ src/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>pts_bomber</title>
 
-    <link data-trunk rel="rust" href="../Cargo.toml" />
+    <link data-trunk rel="rust" href="../Cargo.toml" data-wasm-opt="z" />
     <link data-trunk rel="tailwind-css" href="tailwind.css" />
     <link data-trunk rel="copy-dir" href="public" />
   </head>
